[7.1.0]Return active plugins in plugins.txt order HACK:
Under #4.
Under #5 - notice we need a valid LO passed in for TESV.
ActivePlugins should be a vector but I can't go into this now - wait till
the code is clean.
Note that I also inversed the order of the if for usual case to come first.
This should be a method of game (readPluginsTxt) - see #9.