Add get/setLoadOrder() methods to LoadOrder class - edit:
(ut) fixed mixed tabs and spaces

Setting the load order first checks validity, so that errors are caught
before any changes are made.

The added methods are currently unused, but the LoadOrder class will be
changed to hide how it stores the load order, and all interaction will
need to be done by getters and setters. Because a lot of code currently
interacts directly with the vector methods, I'm implementing these
getters and setters in several steps before making the change.