Hide load order storage mechanism - EEE:
Utumno - omit:

<<<<<<< HEAD
                this->push_back(Plugin(line));
=======

                Plugin plugin(line);
                if (plugin.IsValid(parentGame))
                    loadOrder.push_back(plugin);
>>>>>>> 2e7b8bc... Hide load order storage mechanism

All load order interaction is now done through the getters and setters.
There are a couple of changes in behaviour:

1. lo_fix_plugin_lists() does not remove invalid plugins, because
they're not loaded from files and can't be added in a setter, so the
only way to encounter them is if a plugin is deleted, which would then
trigger the load order to be reloaded anyway.
2. lo_set_load_order() does not add any installed plugins not given to
the end of the load order, because this will be done on the next load
anyway. ---> Utumno - I was already doing this

 # This is the 2nd commit message:

Fixup for parent:

@@ -131,3 +131,3 @@ LIBLO unsigned int lo_set_active_plugins(lo_game_handle gh, const char * const *
     for (const auto& plugin : gh->activePlugins)
-        if (gh->loadOrder.Find(plugin) == gh->loadOrder.end()) {
+        if (gh->loadOrder.getPosition(plugin.Name()) == gh->loadOrder.getLoadOrder().size()) {
             pluginsMissingLO = true;